requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.1
cryptography==41.0.7

# Environment Variables
python-decouple==3.8
# HTTP Requests (for external APIs if needed)
requests==2.31.0

# Utilities
python-dateutil==2.8.2

# Caching (optional, for production)
django-redis==5.4.0
redis==5.0.1

# Task Queue (optional, for async tasks)
# celery==5.3.4
# django-celery-beat==2.5.0

# Monitoring & Logging (optional)
# sentry-sdk==1.39.1

# Development Tools
# django-debug-toolbar==4.2.0
# django-extensions==3.2.3

# Testing
# pytest==7.4.3
# pytest-django==4.7.0
# factory-boy==3.3.0

# Code Quality
# flake8==6.1.0
# black==23.12.1
# isort==5.13.2